requirements.txt
torch==2.3.1
torchvision==0.18.1
Pillow==10.4.0
PyQt5==5.15.11
